feat: refine collection queries and result sorting
